extract lxmf cli compat harness
