+++ crates/apps/lxmf-cli/tests/support/mod.rs
@@ -1,0 +1,1 @@
+pub mod python_lxmd_remote_relay;
